+++ shuup/admin/modules/orders/views/__init__.py
@@ -7,7 +7,7 @@
 # LICENSE file in the root directory of this source tree.
 
 from .detail import OrderDetailView, OrderSetStatusView
-from .edit import OrderEditView
+from .edit import OrderEditView, UpdateAdminCommentView
 from .list import OrderListView
 from .log import NewLogEntryView
 from .payment import OrderCreatePaymentView
@@ -17,5 +17,6 @@ from .shipment import OrderCreateShipmentView, ShipmentDeleteView
 __all__ = [
     "NewLogEntryView", "OrderDetailView", "OrderEditView", "OrderListView",
     "OrderCreatePaymentView", "OrderCreateFullRefundView", "OrderCreateRefundView",
-    "OrderCreateShipmentView", "OrderSetStatusView", "ShipmentDeleteView"
+    "OrderCreateShipmentView", "OrderSetStatusView", "ShipmentDeleteView",
+    "UpdateAdminCommentView"
 ]